pyproject.toml
[tool.poetry]
name = "save-patch"
version = "0.1.0"
description = ""
authors = ["doai"]

[tool.poetry.source]
name = "nexus"
url = "https://nexus.doai.ai/repository/pypi-public/simple"
default = true

[tool.poetry.dependencies]
python = "^3.6.1"
opencv-python = "^4.5.2"
PyYAML = "^5.4.1"
Pillow = "^8.2.0"
pandas = "1.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
